Allow defining API_KEY in code!!
I guess this is unpopular with API libraries but it can be handy if for testing purposes you can place your API key inside the script.  So I've added default parameter to `__init__()` so that no error is thrown if API_KEY isn't provided as in `_get_params()` the global variable `API_KEY` is called in the function namespace, but here if it isn't assigned as environment variable then it will utilize the one defined on `__init__()` if it wasn't defined then well error will throw.